Slight refactoring and make invdes functions available in invdes namespace
